created separate test files for queries and mutations
