+++ tests/test_person.py
@@ -3,7 +3,6 @@ import os
 import unittest
 
 from trompace.exceptions import UnsupportedLanguageException, NotAMimeTypeException
-from trompace.queries import person as person_query
 
 from trompace.mutations import person
 from tests import util
@@ -14,16 +13,6 @@ class TestPerson(unittest.TestCase):
     def setUp(self) -> None:
         super()
         self.data_dir = os.path.join(os.path.dirname(__file__), "data", "person")
-
-    def test_query(self):
-        expected = util.read_file(self.data_dir, "query_person_parameter.txt")
-        created_person = person_query.query_person(identifier="ff59650b-1d47-4ea5-b356-31fddeb48315")
-        self.assertEqual(created_person, expected)
-
-    def test_query_all(self):
-        expected = util.read_file(self.data_dir, "query_person.txt")
-        created_person = person_query.query_person()
-        self.assertEqual(created_person, expected)
 
     def test_create(self):
         expected = util.read_file(self.data_dir, "create_person.txt")
